maint: require pyuvdata>=2.3.1
